refactor(agents-runtime): move skills loader + use_skill tool into runtime
Lifts the `.md`-skill-directory loader and the use_skill / remove_skill
tool builder out of `@electric-ax/agents` and into the leaf
`@electric-ax/agents-runtime` package, alongside the rest of the
entity-runtime primitives (handler context, agent loop, tool plumbing,
low-cost-model completion).

Skills are entity-agnostic: any agents-runtime entity can mount the
tools, and the registry only uses runtime-side dependencies
(`completeWithLowCostModel`, `runtimeLog`). Today they sit in
`@electric-ax/agents` purely for historical reasons — Horton was built
first and the skills code was written next to it.

Concrete moves:

- `src/skills/{registry,tools,preamble,extract-meta,types}.ts` →
  `agents-runtime/src/skills/`. Same files, no behavioural change.
- `serverLog` import in registry.ts and extract-meta.ts replaced with
  the existing `runtimeLog` from `agents-runtime/src/log.ts` (same
  pino sink, same levels, slightly different prefix-as-first-arg call
  signature — adjusted at call sites).
- Public exports added to `agents-runtime/src/index.ts`:
  `createSkillsRegistry`, `createSkillTools`, `SkillsRegistry`,
  `SkillMeta`.
- Existing call sites inside `@electric-ax/agents` (Horton's
  `agents/horton.ts` and the bootstrap) now import from
  `@electric-ax/agents-runtime` instead of the relative path.
- Skill tests move alongside the code:
  `agents-runtime/test/skills/{integration,preamble,tools,registry}.test.ts`.

Verified: skills tests 18/18 in their new home; `@electric-ax/agents`
full suite 38/38 with the new import paths.

External consumers (e.g. OpenFactory's `@electric-ax/discord-bot`) get
the skills primitives from the package they already depend on —
`@electric-ax/agents-runtime` — instead of having to pull in
`@electric-ax/agents` (Horton + Worker + the bootstrap + the
desktop-flavoured server) just to access them.
