+++ packages/agents-runtime/test/skills/integration.test.ts
@@ -2,8 +2,8 @@ import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { createSkillsRegistry } from '../src/skills/registry'
-import { createSkillTools } from '../src/skills/tools'
+import { createSkillsRegistry } from '../../src/skills/registry'
+import { createSkillTools } from '../../src/skills/tools'
 
 let tmpDir: string
