+++ packages/agents-runtime/test/skills/registry.test.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
-import { createSkillsRegistry } from '../src/skills/registry'
+import { createSkillsRegistry } from '../../src/skills/registry'
 
 let tmpDir: string
